+++ englishbackstage/src/main/java/sicnu/sixteam/englishbackstage/service/Impl/UserServiceImpl.java
@@ -4,6 +4,7 @@ package sicnu.sixteam.englishbackstage.service.Impl;
 
 import org.springframework.stereotype.Service;
 
+import org.springframework.web.bind.annotation.RequestMapping;
 import sicnu.sixteam.englishbackstage.mapper.UserMapper;
 import sicnu.sixteam.englishbackstage.model.User;
 import sicnu.sixteam.englishbackstage.service.UserService;
