--- englishbackstage/src/main/java/sicnu/sixteam/englishbackstage/service/LoginService.java
@@ -1,9 +1,0 @@
-package sicnu.sixteam.englishbackstage.service;
-
-import sicnu.sixteam.englishbackstage.model.User;
-
-//用户登录接口
-public interface LoginService {
-    User checkuser(String phonenum, String password);
-}
-
